+++ src/main/java/de/codehat/levelbuilder/listener/TileButtonListener.java
@@ -71,8 +71,15 @@ public final class TileButtonListener implements ActionListener {
         // check for special tile types
         if (tileType == TileType.FOX) {
             askForMovementTypeAndSet(tile);
+        } else if (tileType == TileType.SPEED_POWERUP) {
+            askForSpeedPowerUpDetailsAndSet(tile);
         } else {
             tile.setEnemyMovementType(null); // Reset enemy movement type
+
+            // Reset speed power-up details.
+            tile.setAppearChance(null);
+            tile.setTimeOnField(null);
+            tile.setSpeedIncrease(null);
         }
 
         // update tile and buttons
@@ -81,15 +88,49 @@ public final class TileButtonListener implements ActionListener {
     }
 
     /**
+     * Asks the user for the details of the speed power-up.
+     * E.g the appear chance or the speed increase.
+     *
+     * @param tile the concerning tile
+     */
+    private void askForSpeedPowerUpDetailsAndSet(final Tile tile) {
+
+        // Create all JSpinner.
+        JSpinner sAppearChance = new JSpinner(new SpinnerNumberModel(0.5, 0.01, 1.0, 0.01));
+        JSpinner sTimeOnField = new JSpinner(new SpinnerNumberModel(5, 5, 20, 1));
+        JSpinner sSpeedIncrease = new JSpinner(new SpinnerNumberModel(100, 50, 200, 5));
+
+        // Set their names for the tests (if any).
+        sAppearChance.setName("appearChanceSpinner");
+        sTimeOnField.setName("timeOnFieldSpinner");
+        sSpeedIncrease.setName("speedIncreaseSpinner");
+
+        Object[] message = {
+                "Appear Chance (0.01 - 1.0):", sAppearChance,
+                "Time On Field (in seconds):", sTimeOnField,
+                "Speed Increase (in milliseconds):", sSpeedIncrease
+        };
+
+        JOptionPane.showMessageDialog(this.view,
+                message,
+                "Please give additional information",
+                JOptionPane.QUESTION_MESSAGE);
+
+        tile.setAppearChance((double) sAppearChance.getValue());
+        tile.setTimeOnField((int) sTimeOnField.getValue());
+        tile.setSpeedIncrease((int) sSpeedIncrease.getValue());
+    }
+
+    /**
      * Asks the user, which movement type the tile should get and sets it.
      *
      * @param tile the concerning tile
      */
-    private void askForMovementTypeAndSet(Tile tile) {
+    private void askForMovementTypeAndSet(final Tile tile) {
         JComboBox<String> cbEnemyMovementTypes = new JComboBox<>(EnemyMovementType.TYPES);
         cbEnemyMovementTypes.setName("cbEnemyMovementTypes");
 
-        JOptionPane.showMessageDialog( null,
+        JOptionPane.showMessageDialog( this.view,
                 cbEnemyMovementTypes,
                 "Please select an enemy movement type",
                 JOptionPane.QUESTION_MESSAGE);